Completed Homework - Add and Delete
Added a new view controller for an add note table view that takes a
user input and sends it to the new view controller. I realized that I
tethered the two view controllers in a bad way but I really just don’t
have enough time for doing this assignment totally right and doing work
on my own app! And this works!